Clean code; move and create methods and classes.
